requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
decouple==0.0.7
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.53.1
django-appconf==1.0.5
django-compressor==4.3.1
django-environ==0.10.0
django-sass-processor==1.2.2
django-taggit==4.0.0
idna==3.4
libsass==0.22.0
mercadopago==2.2.0
oauthlib==3.2.2
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
rcssmin==1.1.1
requests==2.28.2
requests-oauthlib==1.3.1
rjsmin==1.2.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
